Renamed Timer to ChessTimer, added new class prototypes.
