+++ src/GUI/ChessPanel.java
@@ -4,7 +4,7 @@ import javax.swing.*;
 import java.awt.geom.Rectangle2D;
 import java.awt.*;
 
-import Game.Player;
+import Game.OldPlayerClass;
 import Game.Pieces.*;
 
 
@@ -14,12 +14,12 @@ import Game.Pieces.*;
 public class ChessPanel extends JPanel {
     private Rectangle2D board;
     private final int boardSize = 600;
-    Player P1, P2;
+    OldPlayerClass P1, P2;
 
     ChessPanel() {
         //make new players
-        P1 = new Player(P2, true);
-        P2 = new Player(P1, false);
+        P1 = new OldPlayerClass(P2, true);
+        P2 = new OldPlayerClass(P1, false);
     }
 
     @Override
@@ -44,8 +44,8 @@ public class ChessPanel extends JPanel {
         drawPieces(P2, g2);
     }
 
-    public void drawPieces(Player player, Graphics2D g2) {
-        for (Piece p : player.pieces) {
+    public void drawPieces(OldPlayerClass oldPlayerClass, Graphics2D g2) {
+        for (Piece p : oldPlayerClass.pieces) {
             p.setPixels(boardSize / 8 * p.returnX(), boardSize / 8 * p.returnY());
             int x = p.getPixelX(), y = p.getPixelY();
             Image img = p.returnPieceImage();
