+++ src/GUI/ToolPanel.java
@@ -1,6 +1,7 @@
 package GUI;
 
-import Game.ThreadTimer.Timer;
+import Game.ChessTimer;
+
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.Graphics;
@@ -22,10 +23,10 @@ public class ToolPanel extends JPanel {
     private final JScrollPane HistoryScroll=new JScrollPane(HistoryList,JScrollPane.VERTICAL_SCROLLBAR_ALWAYS,JScrollPane.HORIZONTAL_SCROLLBAR_ALWAYS);
     private final JLabel Screen1=new JLabel();
     private final JLabel  Screen2=new JLabel();
-    private final JLabel TimDesc1=new JLabel(" Timer 1");
-    private final JLabel TimDesc2=new JLabel(" Timer 2");
-    private Timer Timer1;
-    private Timer Timer2;
+    private final JLabel TimDesc1=new JLabel(" ChessTimer 1");
+    private final JLabel TimDesc2=new JLabel(" ChessTimer 2");
+    private ChessTimer chessTimer1;
+    private ChessTimer chessTimer2;
     
     public ToolPanel() {
         setSize(200,350);
@@ -84,33 +85,33 @@ public class ToolPanel extends JPanel {
         HistoryList.addElemen_tolist(newItem);
     }
     public void change_to_Timer1() {
-        Timer1.start();
-        Timer2.pause();
+        chessTimer1.start();
+        chessTimer2.pause();
     }
     public void change_to_Timer2() {
-        Timer2.start();
-        Timer1.pause();
+        chessTimer2.start();
+        chessTimer1.pause();
     }
     public void stop_timers() {
-        Timer1.pause();
-        Timer2.pause();
+        chessTimer1.pause();
+        chessTimer2.pause();
     }
     
     public void start_Again() {
-        if(Timer1!=null) {
-            Timer1.pause();
-            Timer2.pause();
+        if(chessTimer1 !=null) {
+            chessTimer1.pause();
+            chessTimer2.pause();
         }
         
-        Timer1=new Timer(300);
-        Timer2=new Timer(300);
+        chessTimer1 =new ChessTimer(300);
+        chessTimer2 =new ChessTimer(300);
         
-        Timer1.start();
-        Timer2.start();
-        Timer2.pause();
+        chessTimer1.start();
+        chessTimer2.start();
+        chessTimer2.pause();
         
         HistoryList.clean_list();
-        HistoryList.addElemen_tolist("Player: New Moves");
+        HistoryList.addElemen_tolist("OldPlayerClass: New Moves");
         
     }
     
@@ -139,7 +140,7 @@ class myHistoryList extends JList {
     myHistoryList() {
         this.setBackground(Color.ORANGE);
         setModel(listModel);
-        listModel.addElement("Player: New Moves");
+        listModel.addElement("OldPlayerClass: New Moves");
     }
     public void clean_list() {
         listModel.clear();
